fixed edit form validation
