+++ js/api/api-requests/request-call/download-post-data.mjs
@@ -5,6 +5,7 @@ import { optionWithToken } from "../../api-options/only-auth.mjs";
 import { deleteSessionItem, loadSessionItem } from "../../../storage/session-storage.mjs";
 import { displayPostChangeFeedback } from "../../../display/post/post-feedback/display-post-change-feedback.mjs";
 import { changePostVariables } from "../../api-data/change-post-variables.mjs";
+import { validatedForms } from "../../../form/index.mjs";
 
 export async function downloadPostData(method, action) {
    try {
@@ -25,6 +26,7 @@ export async function downloadPostData(method, action) {
 
       toggleComments(response.ok);
       postDropdownHandler(response.ok);
+      validatedForms(response.ok);
    } catch (error) {
       console.log(error);
    }
