+++ js/handlers/posts/post-dropdown-handler.mjs
@@ -1,7 +1,6 @@
 import { deleteRequest } from "../../api/api-requests/request-type/delete-request.mjs";
-import { updateRequest } from "../../api/api-requests/request-type/update-request.mjs";
-import { validateInputs } from "../../form/validatedInput.mjs";
 import { closeEditPostHandler } from "./close-edit-post-handler.mjs";
+import { editFormHandler } from "./edit-form-handler.mjs";
 
 export function postDropdownHandler(responseOk) {
    if (responseOk) {
@@ -16,26 +15,11 @@ export function postDropdownHandler(responseOk) {
                deleteRequest(dropdownOption);
             }
             if (dropdownOption.innerText === "Edit Post") {
+               const editBtn = dropdownOption;
                const editPostForm = document.querySelector("#edit-post-form");
                editPostForm.classList.replace("d-none", "d-flex");
                closeEditPostHandler(editPostForm);
-
-               editPostForm.addEventListener("submit", (event) => {
-                  event.preventDefault();
-
-                  const editFormSubmit = event.target;
-                  const { tags } = editFormSubmit;
-                  const createPostArray = [];
-                  createPostArray.push(tags);
-                  const isValid = validateInputs(createPostArray);
-
-                  if (isValid) {
-                     console.log("here");
-                     updateRequest(editFormSubmit, dropdownOption);
-                  } else {
-                     console.log(isValid);
-                  }
-               });
+               editFormHandler(editPostForm, editBtn);
             }
          });
       }
